requirements.txt
pymongo
pandas
requests
spotipy
python-dotenv
